pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "openai-client"
version = "0.1.0"
description = "A Python client for the OpenAI API"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "requests>=2.31.0",
    "pydantic>=2.6.0",
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.8"
readme = "README.md"
license = { text = "MIT" }

[project.urls]
Homepage = "https://github.com/yourusername/openai-client"
Repository = "https://github.com/yourusername/openai-client.git" 